Allow fuzzy key matching
Fuzzy key matching means the meta key matches exactly but the key
doesn’t have to match. We might consider using like a Levenshtein
distance algorithm for this in the future, but there are going to be a
relatively small number of keys that won’t match, so I think we’re ok
for now.